+++ webapp/app/.server/db/dto/l2-cancellation-calls.ts
@@ -2,19 +2,19 @@ import { db } from "@/.server/db";
 import { l2CancellationCalls } from "@/.server/db/schema";
 import { type InferInsertModel, eq } from "drizzle-orm";
 
-export function createL2GovernorProposalCall(
+export function createL2CancellationCall(
   data: InferInsertModel<typeof l2CancellationCalls>,
   { tx }: { tx?: typeof db } = {}
 ) {
   return (tx ?? db).insert(l2CancellationCalls).values(data).returning();
 }
 
-export async function getl2GovernorProposalCallsByProposalId(
-  proposalId: number,
-  { tx }: { tx?: typeof db } = {}
-) {
-  return (tx ?? db)
-    .select()
-    .from(l2CancellationCalls)
-    .where(eq(l2CancellationCalls.proposalId, proposalId));
-}
+// export async function getl2GovernorProposalCallsByProposalId(
+//   proposalId: number,
+//   { tx }: { tx?: typeof db } = {}
+// ) {
+//   return (tx ?? db)
+//     .select()
+//     .from(l2CancellationCalls)
+//     .where(eq(l2CancellationCalls.proposalId, proposalId));
+// }
