+++ webapp/app/.server/db/dto/l2-cancellations.ts
@@ -3,14 +3,14 @@ import { db } from "..";
 import { l2CancellationStatusEnum, l2CancellationsTable } from "../schema";
 import { createOrIgnoreRecord, getFirst } from "./utils/common";
 
-export async function createOrIgnoreL2GovernorProposal(
+export async function createOrIgnoreL2Cancellation(
   data: InferInsertModel<typeof l2CancellationsTable>,
   { tx }: { tx: typeof db } = { tx: db }
 ) {
   return createOrIgnoreRecord(l2CancellationsTable, data, { tx }).then(getFirst);
 }
 
-export async function getActiveL2Governors({ tx }: { tx: typeof db } = { tx: db }) {
+export async function getActiveL2Cancellations({ tx }: { tx: typeof db } = { tx: db }) {
   return tx
     .select()
     .from(l2CancellationsTable)
