+++ webapp/app/routes/app/l2-cancellations/new/_route.tsx
@@ -1,7 +1,7 @@
 import {
   createVetoProposalFor,
   getActiveL2Proposals,
-} from "@/.server/service/l2-governor-proposals";
+} from "@/.server/service/l2-cancellations";
 import { hexSchema } from "@/common/basic-schemas";
 import { Button } from "@/components/ui/button";
 import { Form, FormControl, FormField, FormItem } from "@/components/ui/form";
